test coverage and moving code around
